requirements.txt
httpx
groq
deepgram-sdk
streamlit
python-dotenv